move SumBoundary to FabArray
