+++ Src/Base/AMReX_iMultiFab.H
@@ -516,29 +516,6 @@ public:
                          const FabFactory<IArrayBox>& factory = DefaultFabFactory<IArrayBox>()) override;
 #endif
 
-
-    /**
-    * \brief Sum values in overlapped cells.  The destination is limited to valid cells.
-    */
-    void SumBoundary (const Periodicity& period = Periodicity::NonPeriodic());
-
-    void SumBoundary (int scomp, int ncomp, const Periodicity& period = Periodicity::NonPeriodic());
-
-    void SumBoundary_nowait (const Periodicity& period = Periodicity::NonPeriodic());
-
-    void SumBoundary_nowait (int scomp, int ncomp, const Periodicity& period = Periodicity::NonPeriodic());
-
-    /**
-    * \brief Sum values in overlapped cells.  The destination is limited to valid + ngrow cells.
-    */
-    void SumBoundary (int scomp, int ncomp, IntVect const& ngrow,
-                      const Periodicity& period = Periodicity::NonPeriodic());
-
-    void SumBoundary_nowait (int scomp, int ncomp, IntVect const& ngrow,
-                             const Periodicity& period = Periodicity::NonPeriodic());
-
-    void SumBoundary_finish ();
-
     static void Initialize ();
     static void Finalize ();
 };
