+++ Src/Base/AMReX_iMultiFab.cpp
@@ -610,66 +610,4 @@ OwnerMask (FabArrayBase const& mf, const Periodicity& period, const IntVect& ngr
     return p;
 }
 
-void
-iMultiFab::SumBoundary (const Periodicity& period)
-{
-    SumBoundary(0, n_comp, IntVect(0), period);
-}
-
-void
-iMultiFab::SumBoundary (int scomp, int ncomp, const Periodicity& period)
-{
-    SumBoundary(scomp, ncomp, IntVect(0), period);
-}
-
-void
-iMultiFab::SumBoundary (int scomp, int ncomp, IntVect const& nghost, const Periodicity& period)
-{
-    BL_PROFILE("iMultiFab::SumBoundary()");
-
-    SumBoundary_nowait(scomp, ncomp, nghost, period);
-    SumBoundary_finish();
-}
-
-void
-iMultiFab::SumBoundary_nowait (const Periodicity& period)
-{
-    SumBoundary_nowait(0, n_comp, IntVect(0), period);
-}
-
-void
-iMultiFab::SumBoundary_nowait (int scomp, int ncomp, const Periodicity& period)
-{
-    SumBoundary_nowait(scomp, ncomp, IntVect(0), period);
-}
-
-void
-iMultiFab::SumBoundary_nowait (int scomp, int ncomp, IntVect const& nghost, const Periodicity& period)
-{
-    BL_PROFILE("iMultiFab::SumBoundary_nowait()");
-
-    if ( n_grow == IntVect::TheZeroVector() && boxArray().ixType().cellCentered()) return;
-
-    iMultiFab* tmp = new iMultiFab( boxArray(), DistributionMap(), ncomp, n_grow, MFInfo(), Factory() );
-    iMultiFab::Copy(*tmp, *this, scomp, 0, ncomp, n_grow);
-    this->setVal(0.0, scomp, ncomp, nghost);
-    this->ParallelCopy_nowait(*tmp,0,scomp,ncomp,n_grow,nghost,period,FabArrayBase::ADD);
-
-    // All local. Operation complete.
-    if (!this->pcd) { delete tmp; }
-}
-
-void
-iMultiFab::SumBoundary_finish ()
-{
-    BL_PROFILE("iMultiFab::SumBoundary_finish()");
-
-    // If pcd doesn't exist, ParallelCopy was all local and operation was fully completed in "SumBoundary_nowait".
-    if ( (n_grow == IntVect::TheZeroVector() && boxArray().ixType().cellCentered()) || !(this->pcd) ) return;
-
-    FabArray<IArrayBox>* tmp = const_cast<FabArray<IArrayBox>*> (this->pcd->src);
-    this->ParallelCopy_finish();
-    delete tmp;
-}
-
 }
